Add tests for decoder
