+++ server/src/main/scala/lsug/markup/PollenParser.scala
@@ -1,7 +1,6 @@
 package lsug
 package markup
 
-import cats.data._
 import cats.implicits._
 
 import Parser._
@@ -14,7 +13,6 @@ private object PollenParser {
   private val close: Parser[Unit] = pattern("\\}".r).void
   private val name: Parser[String] = pattern("[a-z]+".r)
   private val tagname: Parser[String] = lozenge *> name <* open
-  private val whitespace: Parser[Unit] = pattern("\\s".r).void
   private val contents: Parser[Contents] = pattern("[^◊}]+".r).map(Contents)
 
   def tag: Parser[Tag] =
@@ -24,12 +22,12 @@ private object PollenParser {
     )) { case (name, children) => Tag(name, children)}
 
   private def pollen: Parser[Pollen] = either(contents, tag)
+
   private def children: Parser[List[Pollen]] = zeroOrMore(pollen)
     .map {_.filter {
       case Contents(text) => !text.trim.isEmpty
       case _ => true
    }}
 
-  def tags: Parser[NonEmptyList[Tag]] =
-    oneOrMore(zeroOrMore(whitespace) *> tag <* zeroOrMore(whitespace))
+  def pollens: Parser[List[Pollen]] = children
 }
