+++ server/test/src/scala/lsug/markup/DecoderSpec.scala
@@ -1,112 +1,130 @@
 package lsug
 package markup
 
-import java.time._
-import munit.{Tag => _, _}
+import cats._
 import cats.implicits._
-import cats.data._
 
-class DecoderSpec extends FunSuite {
+class DecoderSpec extends LsugSuite {
 
   import Pollen._
-  import PollenDecoders._
+  import Decoder._
 
-  def assertSuccess[A, B](
-    decoder: Decoder[A, B],
-    input: A,
-    expected: B): Unit = {
-    assert(clue(decoder(input)) == Right(expected))
+  def assertOutput[A](result: A, expected: A)(implicit loc: munit.Location): Unit = {
+    assert(clue(result) == clue(expected), "Result of decoding is incorrect")
   }
 
-  def assertFailure[A, B](
-    decoder: Decoder[A, B],
-    input: A,
-    expected: Decoder.Failure): Unit = {
-    assert(clue(decoder(input)) == Left(expected))
+  def assertEither[A](
+    decoder: Decoder[A],
+    input: List[Pollen],
+    expected: Either[DecoderError, A])(implicit loc: munit.Location): TestBuilder = {
+    val result = decoder(input)
+    expected match {
+      case Right(expected) => (builder {
+          assert(clue(result.isRight), "Failed to decode")
+          result.foreach(assertOutput(_, expected))
+        }).label("success")
+      case Left(expected) => builder({
+        assert(clue(result.isLeft), "Decoding succeeded")
+        result.swap.foreach(assertOutput(_, expected))
+      }).label("fail")
+    }
   }
 
-  test("root tag") {
-    assertSuccess(tag("root"),
-      NonEmptyList.of(Tag("root", Nil)),
-      Tag("root", Nil)
-    )
+  def checkText(name: String)(input: List[Pollen], output: String)(implicit loc: munit.Location): Unit = {
+    assertEither(text, input, Right(output))
+      .label("text")
+      .build(name)
   }
 
-  test("wrong root tag") {
-    assertFailure(tag("root"),
-      NonEmptyList.of(Tag("foo", Nil)),
-      Decoder.Failure(Decoder.Error.TagNotFound("root"), Nil)
-    )
+  def checkTextFails(name: String)(input: List[Pollen], expected: DecoderError)(implicit loc: munit.Location): Unit = {
+      assertEither(text, input, Left(expected))
+        .label("text")
+    .build(name)
   }
 
-
-  test("child tag") {
-    assertSuccess(child("bar"),
-      Tag("foo", List(Tag("bar", Nil))),
-      Tag("bar", Nil)
-    )
+  def checkChildren(name: String)(tagname: String, input: List[Pollen], output: List[Tag]): Unit = {
+    assertEither(children(tagname), input, Right(output))
+      .label("children")
+      .build(name)
   }
 
-  test("wrong child tag") {
-    assertFailure(child("qux"),
-      Tag("foo", List(Tag("bar", Nil))),
-      Decoder.Failure(Decoder.Error.TagNotFound("qux"), Nil)
-    )
+  def checkChild(name: String)(tagname: String, input: List[Pollen], output: Tag): Unit = {
+    assertEither(child(tagname), input, Right(output))
+      .label("child")
+    .build(name)
   }
 
-  test("nested child tag") {
-    assertSuccess(child("bar") >>> child("baz") ,
-      Tag("foo", List(Tag("bar", List(Tag("baz", Nil))))),
-      Tag("baz", Nil)
-    )
+  def checkChildFails(name: String)(tagname: String, input: List[Pollen], expected: DecoderError)(implicit loc: munit.Location): Unit = {
+      assertEither(child(tagname),
+        input,
+        Left(expected))
+        .label("child")
+    .build(name)
   }
 
-  test("wrong nested child tag") {
-    assertFailure(child("bar") >>> child("qux") ,
-      Tag("foo", List(Tag("bar", List(Tag("baz", Nil))))),
-      Decoder.Failure(Decoder.Error.TagNotFound("qux"), List("bar"))
-    )
+  def checkAndThen[A](name: String)(first: Decoder[Tag], second: Decoder[A],
+    input: List[Pollen],
+    output: A): Unit = {
+    assertEither(first.andThen(second), input, Right(output))
+      .label("andThen")
+      .build(name)
   }
 
-  test("wrong parent of nested child tag") {
-    assertFailure(child("bar") >>> child("qux") ,
-      Tag("foo", List(Tag("cow", List(Tag("baz", Nil))))),
-      Decoder.Failure(Decoder.Error.TagNotFound("bar"), Nil)
-    )
+  def checkAndThenTraverse[F[_]: Traverse, A](name: String)(first: Decoder[F[Tag]],
+    second: Decoder[A],
+    input: List[Pollen],
+    output: F[A]): Unit = {
+    assertEither(first.andThenTraverse(second), input, Right(output))
+      .label("andThenTraverse")
+      .build(name)
   }
 
-  test("contents") {
-    assertSuccess(contents,
-      Tag("foo", List(Contents("bar"))),
+  def checkAndThenTraverseFails[F[_]: Traverse, A](name: String)(first: Decoder[F[Tag]],
+    second: Decoder[A],
+    input: List[Pollen],
+    output: DecoderError): Unit = {
+    assertEither(first.andThenTraverse(second), input, Left(output))
+      .label("andThenTraverse")
+      .build(name)
+  }
+
+  checkText("single contents")(List(Contents("hello")), "hello")
+  checkText("multiple contents")(List(Contents("hello"), Contents("world")), "helloworld")
+  checkText("no contents")(Nil, "")
+  checkTextFails("tag")(List(Tag("anything", Nil)), DecoderError.UnexpectedTag("anything"))
+  checkTextFails("tag and contents")(List(Tag("anything", Nil), Contents("hello")), DecoderError.UnexpectedTag("anything"))
+
+  checkChildren("tag")("hello", List(Tag("hello", Nil)), List(Tag("hello", Nil)))
+  checkChildren("no tags")("hello", Nil, Nil)
+  checkChildren("other tags")("hello", List(Tag("hello", Nil), Tag("world", Nil)), List(Tag("hello", Nil)))
+  checkChildren("other tags")("hello", List(Tag("hello", Nil), Tag("world", Nil)), List(Tag("hello", Nil)))
+  checkChildren("contents ignored")("hello", List(Tag("hello", Nil), Contents("world")), List(Tag("hello", Nil)))
+
+  checkChild("tag")("hello", List(Tag("hello", Nil)), Tag("hello", Nil))
+  checkChildFails("multiple tags")("hello", List(Tag("hello", Nil), Tag("hello", Nil)), DecoderError.MultipleTagsFound("hello", 2))
+  checkChildFails("no tags")("hello", Nil, DecoderError.TagNotFound("hello"))
+
+  // // a nested tag
+  checkAndThen("tag containing text")(
+      child("hello"), text,
+      List(Tag("hello", List(Contents("bar")))),
       "bar"
     )
-  }
 
-  test("nel contents") {
-    assertSuccess(contents >>> nel,
-      Tag("foo", List(Contents("bar,baz"))),
-      NonEmptyList.of("bar", "baz")
-    )
-  }
+  checkAndThenTraverse("children containing text")(
+    children("hello"), text,
+    List(Tag("hello", List(Contents("bar"))),
+      Tag("hello", List(Contents("baz")))
+    ),
+    List("bar", "baz")
+  )
 
-  test("date contents") {
-    assertSuccess(contents >>> date,
-      Tag("foo", List(Contents("2000-11-10"))),
-      LocalDate.of(2000, 11, 10)
-    )
-  }
-
-  test("time range") {
-    assertSuccess(contents >>> timeRange,
-      Tag("foo", List(Contents("19:30-20:45"))),
-      LocalTime.of(19, 30) -> LocalTime.of(20, 45)
-    )
-  }
-
-  test("optional tag") {
-    assertSuccess(child("bar").optional.composeF(contents),
-      Tag("foo", Nil),
-      None
-    )
-  }
+  checkAndThenTraverseFails("unexpected tag")(
+    children("hello"),
+    text,
+    List(Tag("hello", List(Tag("bar", Nil))),
+      Tag("world", List(Contents("baz")))
+    ),
+    DecoderError.UnexpectedTag("bar")
+  )
 }
